1.修改msg_group_t中idx和count的类型 2.修改mtu为internal_mtu并去除server端启动时设置mtu的代码

+++ jump/step11/network/msg_group.h
@@ -5,9 +5,9 @@
 
 typedef struct
 {
-    msg_t** elements;
-    size_t  idx;
-    size_t  count;
+    msg_t**        elements;
+    unsigned int   idx;
+    unsigned short count;
 } msg_group_t;
 
 extern void msg_group_free(msg_group_t* g);
